+++ src_cpp/lzbosses.h
@@ -9,7 +9,7 @@ constexpr int amountgearpieces = (6*14);
 class LZBosses {
 	public:
 		LZBosses(int ma_tri);
-		std::vector<int> event(std::vector<int> &all_tries);
+		std::vector<short> event(std::vector<short> &all_tries);
 	private:
 		int max_tries;
 		bool boolinarray(bool* boolarray, bool contains = true);
